+++ torchmetrics/audio/pit.py
@@ -69,6 +69,7 @@ class PIT(Metric):
         Signal Process. ICASSP, IEEE, New Orleans, LA, 2017: pp. 241–245. https://doi.org/10.1109/ICASSP.2017.7952154.
     """
 
+    is_differentiable = True
     sum_pit_metric: Tensor
     total: Tensor
 
@@ -110,7 +111,3 @@ class PIT(Metric):
     def compute(self) -> Tensor:
         """Computes average PIT metric."""
         return self.sum_pit_metric / self.total
-
-    @property
-    def is_differentiable(self) -> bool:
-        return True