+++ torchmetrics/functional/retrieval/reciprocal_rank.py
@@ -18,7 +18,7 @@ from torchmetrics.utilities.checks import _check_retrieval_functional_inputs
 
 
 def retrieval_reciprocal_rank(preds: Tensor, target: Tensor) -> Tensor:
-    """Computes reciprocal rank (for information retrieval).
+    """Computes reciprocal rank (for information retrieval). See `Mean Reciprocal Rank`_
 
     ``preds`` and ``target`` should be of the same shape and live on the same device. If no ``target`` is ``True``,
     0 is returned. ``target`` must be either `bool` or `integers` and ``preds`` must be `float`,
@@ -37,8 +37,6 @@ def retrieval_reciprocal_rank(preds: Tensor, target: Tensor) -> Tensor:
         >>> target = torch.tensor([False, True, False])
         >>> retrieval_reciprocal_rank(preds, target)
         tensor(0.5000)
-
-    .. explained: https://en.wikipedia.org/wiki/Mean_reciprocal_rank
     """
     preds, target = _check_retrieval_functional_inputs(preds, target)