+++ torchmetrics/text/bleu.py
@@ -26,8 +26,7 @@ from torchmetrics.functional.text.bleu import _bleu_score_compute, _bleu_score_u
 
 
 class BLEUScore(Metric):
-    """Calculate `BLEU score <https://en.wikipedia.org/wiki/BLEU>`_ of machine translated text with one or more
-    references.
+    """Calculate `BLEU score`_ of machine translated text with one or more references.
 
     Args:
         n_gram:
@@ -54,12 +53,14 @@ class BLEUScore(Metric):
 
     References:
         [1] BLEU: a Method for Automatic Evaluation of Machine Translation by Papineni,
-        Kishore, Salim Roukos, Todd Ward, and Wei-Jing Zhu http://www.aclweb.org/anthology/P02-1040.pdf
+        Kishore, Salim Roukos, Todd Ward, and Wei-Jing Zhu `BLEU`_
 
         [2] Automatic Evaluation of Machine Translation Quality Using Longest Common Subsequence
-        and Skip-Bigram Statistics by Chin-Yew Lin and Franz Josef Och https://aclanthology.org/P04-1077.pdf
+        and Skip-Bigram Statistics by Chin-Yew Lin and Franz Josef Och `Machine Translation Evolution`_
     """
 
+    is_differentiable = False
+    higher_is_better = True
     trans_len: Tensor
     ref_len: Tensor
     numerator: Tensor
@@ -117,7 +118,3 @@ class BLEUScore(Metric):
         return _bleu_score_compute(
             self.trans_len, self.ref_len, self.numerator, self.denominator, self.n_gram, self.smooth
         )
-
-    @property
-    def is_differentiable(self) -> bool:
-        return False